Add additional documentation for each constructor
